refactor(Search): speed up search perf, filter on client
